requirements.txt
# Core dependencies for ML Systems Evaluation Framework
# Industrial AI Reliability Assessment

# Core framework dependencies
PyYAML>=6.0.1
requests>=2.31.0

# Configuration and validation
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.1.0

# CLI and user interface
click>=8.1.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Type checking
mypy>=1.5.0
types-PyYAML>=6.0.12
types-requests>=2.31.0

# Code quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
